Move CA configmap code to a separate file
* pkg/operator/controller/controller.go (caCertConfigMapName):
(GlobalMachineSpecifiedConfigNamespace, shouldPublishRouterCA):
(ensureRouterCAIsPublished, ensureRouterCAIsUnpublished): Move from here...
* pkg/operator/controller/controller_ca_configmap.go: ...to this new file.
